make screen with float32
make screen with float32 dtype as a start. Somewhere later it still seems to change to float64 because the final rbf fits file has bitpix=-64. Still need to find out where/how that happens. That would save memory consumption...which is already a major issue.

+++ utils/make_gain_screen.py
@@ -57,7 +57,7 @@ def interpolate_station(antname, ifstep):
     xx, yy = np.meshgrid(x, y)
     # Do radial basis function interpolation for each time step.
     # First axis of data is time.
-    screen = np.ones((data.shape[0], 1, 1, 4, 256, 256))
+    screen = np.ones((data.shape[0], 1, 1, 4, 256, 256), dtype=np.float32)
     X, Y = np.around(wcs.wcs_world2pix(RA, DEC, 0)).astype(int)
     # data has shape (time, freq, ant, matrix, y, x)
     # gains has shape (time, freq, ant, dir, pol)
